Solves issue #1233
