+++ plugins/de.ovgu.featureide.fm.ui/src/de/ovgu/featureide/fm/ui/editors/featuremodel/operations/ElementDeleteOperation.java
@@ -28,19 +28,15 @@ import java.util.Collections;
 import java.util.List;
 
 import org.eclipse.gef.ui.parts.GraphicalViewerImpl;
-import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.viewers.TreeViewer;
 
-import de.ovgu.featureide.fm.core.base.FeatureUtils;
 import de.ovgu.featureide.fm.core.base.IConstraint;
 import de.ovgu.featureide.fm.core.base.IFeature;
 import de.ovgu.featureide.fm.core.base.IFeatureModel;
-import de.ovgu.featureide.fm.core.base.IFeatureStructure;
 import de.ovgu.featureide.fm.core.io.manager.FeatureModelManager;
 import de.ovgu.featureide.fm.core.io.manager.IFeatureModelManager;
 import de.ovgu.featureide.fm.core.localization.StringTable;
-import de.ovgu.featureide.fm.ui.editors.DeleteDialog;
 import de.ovgu.featureide.fm.ui.editors.FeatureModelEditor;
 import de.ovgu.featureide.fm.ui.editors.featuremodel.GUIDefaults;
 import de.ovgu.featureide.fm.ui.editors.featuremodel.editparts.ConstraintEditPart;
@@ -61,9 +57,12 @@ public class ElementDeleteOperation extends MultiFeatureModelOperation implement
 
 	private final Object viewer;
 
-	public ElementDeleteOperation(Object viewer, IFeatureModelManager featureModelManager) {
+	private final String dialogReturnLabel;
+
+	public ElementDeleteOperation(Object viewer, IFeatureModelManager featureModelManager, String dialogReturnLabel) {
 		super(featureModelManager, DELETE, getFeatureNames(viewer));
 		this.viewer = viewer;
+		this.dialogReturnLabel = dialogReturnLabel;
 	}
 
 	@Override
@@ -128,70 +127,13 @@ public class ElementDeleteOperation extends MultiFeatureModelOperation implement
 			}
 		}
 
-		// Check conditions for slicing
-		boolean featureInConstraint = false;
-		boolean featureHasGroupDifference = false;
-		boolean featureIsRoot = false;
-		for (final IFeature feature : featuresToDelete) {
-			if (!FeatureUtils.getRelevantConstraints(feature).isEmpty()) {
-				featureInConstraint = true;
-			}
-			if (hasGroupDifference(feature)) {
-				featureHasGroupDifference = true;
-			}
-			if (isRootRequiringSlicing(feature, featuresToDelete)) {
-				featureIsRoot = true;
-			}
-		}
-
-		if (featureInConstraint || featureHasGroupDifference || featureIsRoot) {
-			// the delete dialog needs to be shown
-			final List<String> dialogReasons = getDialogReasons(featureInConstraint, featureHasGroupDifference, featureIsRoot);
-			final String[] dialogButtonLabels = getDialogButtonLabels(featureInConstraint, featureHasGroupDifference, featureIsRoot);
-			final String dialogReturnLabel = openDeleteDialog(featuresToDelete.size() > 1, dialogReasons, dialogButtonLabels);
+		if (dialogReturnLabel != null) {
 			handleDialogReturn(dialogReturnLabel, featuresToDelete, constraintsToDelete);
 		} else {
 			// regular delete
 			addDeleteConstraintOperations(constraintsToDelete);
 			addDeleteFeatureOperations(featuresToDelete);
 		}
-	}
-
-	/**
-	 * Checks if the group of the feature and the group of its parent are of the same type
-	 *
-	 * @param feature The feature of which the group is compared to the group of its parent
-	 * @return <code>true</code> if the group of the feature is different from the group of the parent. <code>false</code> if the group is the same or the
-	 *         feature has no parent or no children.
-	 */
-	private boolean hasGroupDifference(IFeature feature) {
-		final IFeature parent = FeatureUtils.getParent(feature);
-		if ((parent == null) || !FeatureUtils.hasChildren(feature)) {
-			return false;
-		}
-		return (FeatureUtils.isOr(feature) && !FeatureUtils.isOr(parent)) || (FeatureUtils.isAlternative(feature) && !FeatureUtils.isAlternative(parent))
-			|| (FeatureUtils.isAnd(feature) && !FeatureUtils.isAnd(parent));
-	}
-
-	/**
-	 * Checks whether the given feature is the root feature, and deleting it requires slicing because there is no unambiguous new root.
-	 *
-	 * @param feature The feature to check
-	 * @param featuresToDelete The list of features to be deleted
-	 * @return <code>true</code> iff the given feature is the root, and deleting it requires slicing.
-	 */
-	private boolean isRootRequiringSlicing(IFeature feature, List<IFeature> featuresToDelete) {
-		if (!FeatureUtils.isRoot(feature)) {
-			return false;
-		}
-		IFeatureStructure fs = feature.getStructure();
-		while (fs.getChildrenCount() == 1) {
-			fs = fs.getFirstChild();
-			if (!featuresToDelete.contains(fs.getFeature())) {
-				return false;
-			}
-		}
-		return true;
 	}
 
 	/**
@@ -209,70 +151,6 @@ public class ElementDeleteOperation extends MultiFeatureModelOperation implement
 			addDeleteConstraintOperations(constraintsToDelete);
 			addDeleteFeatureOperations(featuresToDelete);
 		}
-	}
-
-	/**
-	 * Opens the DeleteDialog with the given options
-	 *
-	 * @param multiple <code>true</code> if multiple features are being deleted, <code>false</code> if not
-	 * @param dialogReasons A List of Strings with reasons for the DeleteDialog. These are being displayed in the dialog
-	 * @param dialogButtonLabels A String array with labels for the buttons of the DeleteDialog
-	 * @return A String containing the label of the button that was pressed in the DeleteDialog or <code>null</code> if the dialog was closed differently
-	 */
-	protected String openDeleteDialog(boolean multiple, List<String> dialogReasons, String[] dialogButtonLabels) {
-		final MessageDialog dialog = new DeleteDialog(null, multiple, dialogReasons, dialogButtonLabels, dialogButtonLabels.length - 1);
-		dialog.open();
-		final int dialogReturn = dialog.getReturnCode();
-
-		if ((dialogReturn >= 0) && (dialogReturn < dialogButtonLabels.length)) {
-			return dialogButtonLabels[dialogReturn];
-		} else {
-			return null;
-		}
-	}
-
-	/**
-	 * Gets the button labels for the DeleteDialog
-	 *
-	 * @param featureInConstraint <code>true</code> if any of the selected features is contained in a constraint, <code>false</code> if not
-	 * @param featureHasGroupDifference <code>true</code> if any of the selected features has a different group than its parent, <code>false</code> if not
-	 * @param featureIsRoot <code>true</code> if any of the selected features is the root and has multiple children, <code>false</code> if not
-	 * @return A String array with labels for the buttons of the DeleteDialog
-	 */
-	private String[] getDialogButtonLabels(boolean featureInConstraint, boolean featureHasGroupDifference, boolean featureIsRoot) {
-		final List<String> buttonLabels = new ArrayList<>();
-		if (featureInConstraint || featureHasGroupDifference || featureIsRoot) {
-			buttonLabels.add(StringTable.DELETE_WITH_SLICING);
-		}
-
-		if (!featureInConstraint && !featureIsRoot) {
-			buttonLabels.add(StringTable.DELETE_WITHOUT_SLICING);
-		}
-
-		buttonLabels.add(StringTable.CANCEL);
-		return buttonLabels.toArray(new String[0]);
-	}
-
-	/**
-	 * Gets the reasons for the DeleteDialog
-	 *
-	 * @param featureInConstraint <code>true</code> if any of the selected features is contained in a constraint, <code>false</code> if not
-	 * @param featureHasGroupDifference <code>true</code> if any of the selected features has a different group than its parent, <code>false</code> if not
-	 * @param featureIsRoot <code>true</code> if any of the selected features is the root and has multiple children, <code>false</code> if not
-	 * @return A List of Strings with reasons for the dialog. These are being displayed in the DeleteDialog
-	 */
-	private List<String> getDialogReasons(boolean featureInConstraint, boolean featureHasGroupDifference, boolean featureIsRoot) {
-		final List<String> dialogReasons = new ArrayList<>();
-		if (featureInConstraint) {
-			dialogReasons.add(StringTable.DELETE_FEATURE_REASON_CONSTRAINTS);
-		}
-		if (featureHasGroupDifference) {
-			dialogReasons.add(StringTable.DELETE_FEATURE_REASON_GROUP_DIFFERENCE);
-		}
-		if (featureIsRoot) {
-			dialogReasons.add(StringTable.DELETE_FEATURE_REASON_ROOT);
-		}
-		return dialogReasons;
 	}
 
 	/**
