+++ plugins/de.ovgu.featureide.fm.ui/src/de/ovgu/featureide/fm/ui/views/constraintview/actions/DeleteConstraintInViewAction.java
@@ -55,7 +55,7 @@ public class DeleteConstraintInViewAction extends AbstractConstraintEditorAction
 		if (selection.toList().size() == 1) {
 			FeatureModelOperationWrapper.run(new DeleteConstraintOperation((IConstraint) selection.getFirstElement(), featureModelManager));
 		} else if (selection.toList().size() > 1) {
-			FeatureModelOperationWrapper.run(new ElementDeleteOperation(viewer, featureModelManager));
+			FeatureModelOperationWrapper.run(new ElementDeleteOperation(viewer, featureModelManager, null));
 		}
 	}
